权限

+++ application/index/controller/Logs.php
@@ -23,10 +23,15 @@ class Logs extends Base
     }
 
     public function ajax_delete_logs(){
-        $day = Request::instance()->get("day");
-        $time = date("Y-m-d H:i:s",time() - $day*3600*24);
-
-       return model("log","logic")->delete_logs_by_time($time);
+        if(!$this->check_post_menu_permission("delete_operate")){
+            return '没有权限！';
+        }
+        else {
+            $day = Request::instance()->get("day");
+            $time = date("Y-m-d H:i:s", time() - $day * 3600 * 24);
+            model("log", "logic")->delete_logs_by_time($time);
+            return  0;
+        }
     }
