+++ application/index/controller/User.php
@@ -15,17 +15,30 @@ class User extends Base{
 
 
     public  function  save_user(){
-        $data = Request::instance()->post();
-        $file = Request::instance()->file("heard_img");
-        $uid = model('user','logic')->save_user($data,$file);
-        $this->redirect("system/user_des",["id"=>$uid]);
+
+        if(!$this->check_post_menu_permission("update_operate")){
+            echo "<script> alert('没有权限！');history.back(-1);</script>";
+        }
+       else{
+           $data = Request::instance()->post();
+           $file = Request::instance()->file("heard_img");
+           $uid = model('user','logic')->save_user($data,$file);
+           $this->redirect("system/user_des",["id"=>$uid]);
+       }
+
     }
 
 
     public  function user_del($id){
 
-        model("user","logic")->delete_user($id);
-        $this->redirect("system/member");
+
+        if(!$this->check_post_menu_permission("delete_operate")){
+            echo "<script> alert('没有权限！');history.back(-1);</script>";
+        }
+        else {
+            model("user", "logic")->delete_user($id);
+            $this->redirect("system/member");
+        }
     }
